contrib: Add Dockerfile to build and configure a FreeSWITCH
Rhizomatica is using FreeSWITCH and we should have an easy way to
test against it. A docker container with exposed ports seems like
the easiest. FreeSWITCH by default is giving us some exmaple numbers:

	* 5000 a menu... that allows DTMF
	* 9195 an echo test
	* 9198 tetris.

The config is copied on top of the default/big config that is
installed. If this PBX should be reached from the outside one needs
to change 127.0.0.1 to the external address and maybe configure the
acl as well to add more CIDRs.

Besides that
	make container
	make run

Will build it and start the container. Takes a bit of time and requires
docker. With it configure one can see things like:

2017-03-05 15:32:49.913912 [INFO] switch_channel.c:515 RECV DTMF 3:2000
2017-03-05 15:32:50.952752 [INFO] switch_channel.c:515 RECV DTMF 2:2000

Now to test DTMF in the system.

Change-Id: I7f3aa8c81b9e8698df090a05d2e41a41b67d8e3c